Moved mono icons to an extra package next to the License.md
